hw2/requirements.txt
flask==1.1.2
requests2==2.16.0
Pillow==2.2.2

